Define /v1/auth/authorization_code_url endpoint
This commit add implementation that makes requests to [1], to
respond with the authorization code url so that the frontend app
can use it to compose a login button.

This includes a very rough definition of a routes which will be sorted
later.

We found that the relative paths began with app and app/__init__.py is
modular.

[1] /v1/auth/authorization_code_url