Define endpoint for authenticating users
This commit defines the /v1/auth/callback endpoint which is responsible
for verifying that the aithorization and redirect uri are valid.

It makes an access token request to the IdP which responds valid when
the aforementioned parameters are valid.

An access token is returned, after which the endpoint request for the
user's info.

Upon receipt of the access token and user info a tokenized user is
composed and the return to the frontend.

Tests have been included within that further elaborate.