Add tests for expired token
This tests for when the id_token is expired.. From the tests you will
notice that the UnprocessableEntity is being caught and thrown.

This is becuase Flask restplus appears to have an issue with the
standard flask @api.errorhandler callback.

See the following for more details on this issue:

- https://github.com/noirbizarre/flask-restplus/issues/744
- https://github.com/noirbizarre/flask-restplus/issues/530

The exceptions class is gotten from werkzeug.exceptions[1], help
formulate a JSON response[1], which is expected and checked in the
tests.

[1] https://werkzeug.palletsprojects.com/en/1.0.x/exceptions/