Create project
This commit adds implementation that creates a project and returns
a slug belonging to the project.

This is also the very first feature that requires authentication.

A lot of the tests have been used to ensure that authentication works

The email of the user is used to link to each project