Upload and associate raw file with project
This commit add implementation and tests that enable the upload and
association of a raw file with a project.

We have had to introduce the following Environment variables:

- RAW_FILES_UPLOAD_FOLDER (points to the upload folder)
- ALLOWED_RAW_FILE_EXTENSIONS (list of file extensions)

When a researcher intends on uploading a raw file to associate it with a
project, from the backend `PUT /v1/projects/<slug>/raw_file` is used for
this.

The file is checked for its presents, name and file extension, any of
these that are not provided will result in a 400 error response.

The only allowed file extension is mzXML.

A 404 not found response could ensue if the the supplied lug does not
belong to a project in the database.

Migrations with project to raw files associations and attributes have been
defined within.

Tests also exist