FIX: Change the datatype for access token
This commit changes the datatype for access token from STRING(256)
to TEXT.

This because the datatype of STRING(256) was not able to take that many
characters from the access token supplied by the IdP.

However, TEXT datatype did provide enough space, hence why we have used
it.

Within one would notice that we had to remove the column and add is back
becuase changing the column type alone was not picked up by Alembic.