FEATURE: Allow user to log out
In routes.py added conditional to check if endpoint is /callback or
/authorization_code_url and return on either.
In auth_controller.py: added logout endpoint and AuthLogout class with
delete function to remove access_token from Session table and
set token_user to None.
In session.py we imported the dataclass object to define datatype for
tokenized_user.
In test_user_auth_via_oidc.py the test case for TestUserLogout was added.
This includes the setUp method and test_logout method.