Feature: ensure unique slugs are generated
This commit adds implementation code that ensures that onlt unique slugs
are generated.

This uses a Slugifier class to carry this out, it take 2 parameters
a queryable object and the title, name or any string that requires
a slug.

These slugs are required to be unique so the strategy here is to query
the database slugs that are list this in descending order by id,
returning the first (which is last).

The slug associated with the record is then returned and split by '-'.

It is expected that the last value would be a string or integer.

When a string is returned 1 is appended to the new slug and returned
else if an integer is returned then the the current count is increased
and appended to the new slug.