HOT-FIX: Alter filter to include project collaboarations
This updates the fetch project query to include collaboration
projects belonging to the current user.

This was unintentionally omitted and now included to ensure
the current user get access to projects that require thier attention.

This has been done using SQLAlchemy's SQL or_ wrapper which get
translated to an OP HWERE clause.