HOT-FIX: Return 401 for empty authorization header
This commit fixed the IndexError that is fired when the Authorization
and/or X-ACCESS-TOKEN is passed empty or in the case of the
Authorization bearer just with the 'Bearer'.

This commit ensures that these scenario returns a 401 unauthorized http
status.

This will ensure that the safety and intergrity of the API