Add support for uploading multiple RawFiles
This implementation within this commit ensures that whenever one or more
RawFiles are uploaded via these endpoints (i.e [1] and [2]), they are:

- Uploade and save in a pre-defined remote directory
- Associate each to the supplied project, identified via its slug
- Checksums created for easier changes traceability.

This implementation isn't at the moment optimal enough as it utilizes
one for loop to build through the project_raw_files and another
defined with the ProjectRawFileDAO to iterate through the built
objects and persist.

Due to the nature of this implementation and with the team's growing
knowledge in Python improving. This area has been identified for
future improvements with python iterable generators top on the
list of options to explore, especially contextmanager [3].

This we found to only iterate once for yield returning a weird RuntimeError: generator didn't yield.

These we have identified and will revisit.

Also for the database transactions, we intend to fully making use of bulk_save_objects [4].

[1] /v1/projects/<slug>/raw_file
[2] /v1/projects/<slug>/raw_files
[3] https://docs.python.org/3/library/contextlib.html
[4]  https://docs.sqlalchemy.org/en/13/orm/persistence_techniques.html#bulk-operations