Implement pagination of Projects
This commit adds implementation that paginates projects.

This has been done with the help of flask_sqlalchemy's pagination
[1] implementation.

This implementation appears to be built to cater for frontend templates,
hence why a Pagination object is returned.. With this object [1] comes a
other iterable methods and query property which contains the actual query.

This query isn't serializable into JSON hence why the __dict__ is used
to serialize the Pagination object to a Dictionary and then the tap
method is used to remove the query and then deserializes it to a
ProjectCollection which the DAO caller can choose what to do with it.

[1] https://flask-sqlalchemy.palletsprojects.com/en/2.x/api/#flask_sqlalchemy.Pagination