[TRELLO-63] Import of Metadata Shipment files
This commit implements an import facility for importing or ingest Metadata shipment files that are excel based.

With a xlsx file extension they provide metadata about the samples and are linked to blood isolate number.
The excel file come with the following columns
- Columns:
    - DATE shipped :date
    - MATRIX_BOX : string (needs to be serialize)
    - MATRIX_LOCN: string but comma seperated like A,1.. Needs to be split apart to 2 columns: WELL_ROW and WELL_COL
    - ORGM:string should be renamed from org to ORGANISM.
    - ISOLATE_NBR: ISOLATE_NBR.. identifies the isolates.. in some case it is a blood isolate and in other cases it just an identifier. This file characterises the APL plate (Alberta Public Laboratories)

These we have not made configurable becuase at the time of this commit it was determined that it was very unlikely to change

These columns are used to extract the value belonging to each cell. These are the used to compose a JSON attribute for the metadata shipment row.

The metadata shipments are the returned to the requesting client.

The for loops and the insertion into the database are areas that require improvement. We have marked this for furture refactoring