FIX - change naming convention and references of raw files to mzxml files.
This commit addresses the issue that everywhere in the API which
references raw files needs to be changed to mzxml files. In this
commit all references have been changed, the API runs and passes
all tests, however the Database names/columns all need to be
changed to reflect the changes in the python code. This will enable
the front end to be able to communicate properly.

If a user has already installed the app before this commit, they
will have to drop the db and remake it.

Steps:
1.) Start docker container
  $ docker-compose up

2.) In another terminal tab run these commands:
  $ docker exec -it api /bin/bash
  $ source ./env-packages/bin/activate
    $ make dropdb
    $ make createdb
    $ python manage.py db upgrade head