FIXUP: Stop duplicate tests from running
This creates a class for project tests TestProjectBase

This is used to replace TestCreateProject which was being inherited
by other test classes.

It happens that the tests defined within TestCreateProject get call
before the tests of the child class.

by replacing with TestProjectBase whic has not tests defined.. the
duplicate tests will no longer run.