[TRELLO-61] Performance improvement of builder loop
This commit make changes that help improve the performance and
optimization of the loop.

This changes the ProjectMZXmlFilesBuilder to a iterable object with the
implementation of the __iter__ method and an iterator with the addition
of the __next__.

These dunder methods essential eliminate the need for an extra for loop
in the builder.

All that is required with this setup is tracking and update of the
counter's state, which helps the dunder next method figure out the
next project raw file to make available to the outer for loop.