Associate MzXml files to DataTypes
This commit follows through from the last commit to assoicate MzXml files
with the DataTypes.

This has led to the removal of the database relationship between projects
and MzXml files tables respectively and is followed by its assoicattion
with DataTypes table.

This also necessitates the refactoring of the former endpoint for creating
MzXml files

This changed from '/v1/projects/<slug>/mzxml_files' to '/v1/projects/<slug>/data_types/<data_type_slug>/mzxml_files'

As a result the implementation changes with the DataType now used as the focal point.

Also, changes were change to the tests to this effect with a new FactoryBoy factory
defined that links projects data_types and mzxml_files.