Associate Metadata Shipment files to data_type
This commit follows through from the last commit to assoicate Metadata Shipment files
with the DataTypes.

This has led to the removal of the database relationship between MzXML
and Metadata Shipment files tables respectively and is followed by its assoication
with DataTypes table.

This also necessitates the refactoring of the former endpoint for creating
Metadata Shipment files

This changed from '/v1/projects/<slug>/mzxml_files/<mzxml_files_id>/metadata_shipments' to '/v1/projects/<slug>/data_types/<data_type_slug>/metadata_shipments'

As a result the implementation changes with the DataType now used as the focal point.

Also, changes were made to the tests to this effect with a new FactoryBoy factory
defined for MetadataShipmentFactory