[TRELLO-76] Polymorphic Single Table Inheritance
This commit defines a data_formats table and uses it to organise
the MZXML and metadata shipment files.

This has been done due to the similarities between the aforementioned file data_formats

This also provide us with a single table inheritance model that will be central and easy to\
query based on the type which is all internally handled by SQL Alchemy.

MZXML and metadata shipment tables have been dropped in favour of data_formats table

Also the file_name has been renamed to name.

The tests have been modified in order to reflect the aforementioned changes.