Refactor: Remove Flask-Script and Define seed command
This commit is aimed at removing the Flask-Script manager in favour of
Flask-CLI which already comes bundled with Flask.

This has become necessary becuase of the deprecation warnings that we
keep getting.

With this refactor we have gotten rid of the warning that were
attributed to Flask-SCript.

This came at a huge cost, as we had to make a lot of changes to the
terminal script

A lot of the Flask-Migrate scripts work off of the flask db command

We had to defined and add tests to cover dbL create and drop.

This are the custom script that belong to this project.

Also this affects the way flask is started, as a result FLASK_APP  has
to be set point to manage.py.

This help it to identify the main script.

We have also defined within a seed database task which we hope will
reduce a lot of the time take to create dummy data.

[1] https://blog.miguelgrinberg.com/post/migrating-from-flask-script-to-the-new-flask-cli