Replace Slugifier class with python-slugify
This commit makes sweeping changes that affects the generation
of slugs for Project and DataType respectively.

This has come about as a result of creation of a new slug when any
other attribute's value has changed.

This is behaiviour is confusing, so within Slugifier is being replaced with
python-slugify.

Also, a listener is used to centrally reuse/share the slugify method.

In order to ensure uniqueness of each generated slug, a UUID is appended to it.

UUID[1] are unique and will help avoid conflicts.

Also ids for Project, DataType and DataFormatFiles have been removed.

This has been done because slug already serves its purpose.

[1] https://docs.python.org/2/library/uuid.html#:~:text=If%20node%20is%20not%20given,bit%20sequence%20number%20is%20chosen.&text=Generate%20a%20UUID%20based%20on,(which%20is%20a%20string).&text=Generate%20a%20random%20UUID.